+++ src/viat/providers/storage/__init__.py
@@ -1,7 +1,7 @@
 """A convenience module that lists all [`ViatAttributeStorage`][viat.protocols.ViatAttributeStorage] providers."""
 
 from .json import JsonAttributeReader, JsonAttributeStorage, JsonAttributeStorageConfig, JsonAttributeStorageConnection
-from .toml import TomlAttributeReader, TomlAttributeStorage, TomlAttributeStorageConfig, TomlAttributeStorageConnection
+from .toml import TomlAttributeStorage, TomlAttributeStorageConfig
 
 
 __all__ = [
@@ -10,9 +10,6 @@ __all__ = [
     'JsonAttributeStorageConfig',
     'JsonAttributeStorageConnection',
     'JsonAttributeStorageMutator',
-    'TomlAttributeReader',
     'TomlAttributeStorage',
     'TomlAttributeStorageConfig',
-    'TomlAttributeStorageConnection',
-    'TomlAttributeStorageMutator',
 ]