+++ src/viat/providers/storage/_json/__init__.py
@@ -1,4 +1,5 @@
 from .config import JsonAttributeStorageConfig
 from .connection import JsonAttributeStorageConnection
 from .mutator import JsonAttributeMutator
-from .storage import JsonAttributeStorage
+from .reader import JsonAttributeReader
+from .storage import AbstractJsonAttributeStorage, JsonAttributeStorage