+++ src/viat/providers/storage/json.py
@@ -1,13 +1,17 @@
 """The JSON [`ViatAttributeStorage`][viat.protocols.ViatAttributeStorage] provider."""
 
-from ._json.config import JsonAttributeStorageConfig
-from ._json.connection import JsonAttributeStorageConnection
-from ._json.mutator import JsonAttributeMutator
-from ._json.reader import JsonAttributeReader
-from ._json.storage import JsonAttributeStorage
+from ._json import (
+    AbstractJsonAttributeStorage,
+    JsonAttributeMutator,
+    JsonAttributeReader,
+    JsonAttributeStorage,
+    JsonAttributeStorageConfig,
+    JsonAttributeStorageConnection,
+)
 
 
 __all__ = [
+    'AbstractJsonAttributeStorage',
     'JsonAttributeMutator',
     'JsonAttributeReader',
     'JsonAttributeStorage',