+++ src/viat/providers/storage/toml.py
@@ -1,18 +1,12 @@
-"""The TOML [`ViatAttributeStorage`][viat.protocols.ViatAttributeStorage] provider."""
+"""The TOML [`ViatAttributeStorage`][viat.protocols.ViatAttributeStorage] provider, based on the JSON provider."""
 
 from ._toml import (
-    TomlAttributeMutator,
-    TomlAttributeReader,
     TomlAttributeStorage,
     TomlAttributeStorageConfig,
-    TomlAttributeStorageConnection,
 )
 
 
 __all__ = [
-    'TomlAttributeMutator',
-    'TomlAttributeReader',
     'TomlAttributeStorage',
     'TomlAttributeStorageConfig',
-    'TomlAttributeStorageConnection',
 ]